Initial pass at documentation site (#38)
Uses [vitepress](https://vitepress.dev/), which is simple and powerful.

I went with default themes for now, and reused README content. We can
add guides / programmatically import reference code one it's up.

I plan on deploying them with Vercel to start, but it's a static site so
we have many easy options.