Add StreamToNdJson() function in axflow/models (#61)
This library will encourage [newline-delimited JSON](http://ndjson.org)
as the "standard" interface for streaming data from API endpoints. The
popular model APIs all stream chunks a little differently. Converting
each stream to ND JSON will allow us to build reusable client utilities,
e.g., a `useChat` react hook on the frontend.

For example, say we have the following list of objects that we're going
to stream:

```ts
const stream = createReadableStreamFrom([
  { content: "Some" },
  { content: " content" },
  { content: " goes" },
  { content: " here" },
  { content: "." },
]);
```

When we convert this stream to ND JSON using the `NdJsonStream` utility,
we get:

```ts
for (const chunk of NdJsonStream.from(stream)) {
  console.log(new TextDecoder().decode(chunk));
}
```

```txt
{"type":"chunk","value":{"content":"Some"}}\n
{"type":"chunk","value":{"content":" content"}}\n
{"type":"chunk","value":{"content":" goes"}}\n
{"type":"chunk","value":{"content":" here"}}\n
{"type":"chunk","value":{"content":"."}}\n
```

This format ^ is what we'll encourage for our streaming representation.

If you want to prepend additional data to the stream, you can supply an
object or a list of objects to the `NdJsonStream.from` method as the
second argument:

```ts
for (const chunk of NdJsonStream.from(stream, [{ additional: "data" }, { goes: "here" }])) {
  console.log(new TextDecoder().decode(chunk));
}
```

```txt
{"type":"data","value":{"additional":"data"}}\n
{"type":"data","value":{"goes":"here"}}\n
{"type":"chunk","value":{"content":"Some"}}\n
{"type":"chunk","value":{"content":" content"}}\n
{"type":"chunk","value":{"content":" goes"}}\n
{"type":"chunk","value":{"content":" here"}}\n
{"type":"chunk","value":{"content":"."}}\n
```

Lastly, the `NdJsonStream` has a `headers` property defined on it with
the proper HTTP response headers. This can be convenient when creating
API routes. For example:

```ts
function POST(request: NextRequest) {
  const chatRequest = await request.json();

  const stream = await OpenAIChat.stream(chatRequest, {
    apiKey: process.env.OPENAI_API_KEY!,
  });

  return new Response(NdJsonStream.from(stream), { headers: NdJsonStream.headers });
}
```