Remove stream map functionality for now (#66)
This was some sugar that I am removing for now. I think it's really
useful, but I think it needs some extra consideration, so punting on it
for now.

In particular, what I want out of the response stream mapping would be:

1. Async + Sync mapping (already the case in this implementation that
I'm removing)
2. I want to filter out some values. Some of the streams return entries
without tokens (e.g., openai returns a chunk with a finish reason and no
tokens, I want to filter this out from the token stream).
3. I want to be able to enqueue arbitrary data in addition to the stream
chunks, per out streaming interface.
4. I *may* want additional error handling capabilities (but this might
be handled by returning a rejected promise, so not sure).

Something that looks close to this:

```ts
// API response
return new StreamingJsonResponse(stream, {
  // We want to support async / sync mapping
  async map(chunk, controller) {
    // We may want to filter some chunks out.
    if (!chunk.someProperty) {
      return;
    }

    try {
      const {transformedChunk, someMetadata} = await transformChunkAsync(chunk);
      // Enqueue the chunk like a normal stream controller enqueue call
      controller.enqueue(transformedChunk);
      // Send additional data / metadata to the client alongside the stream
      controller.enqueueData(someMetadata);
    } catch (error) {
      // Some errors may be problematic enough that we want to error the stream. Others we can ignore.
      if (error.fatal) {
        controller.error(error);
      }
    }
  }
});
```

This needs more though, so removing for now and will come back to this
later.